requirements.txt
qrcode[pil]==7.4.2
Pillow>=9.0.0